Team-project-1-semester-main/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
typing_extensions==4.9.0
yarl==1.9.4
